New Function
I added a new function for the Nationalities table. Hopefully I will learn and setup joins in the functions to make it all look a lot nicer. I experimented with adding a way of writing your own queries as the user, but I haven't got it working to a satisfactory level so I just removed it for now.